# micro/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/config/configKey/{key}:
        get:
            tags:
                - SysConfig
            operationId: SysConfig_ConfigByKey
            parameters:
                - name: key
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.ConfigByKeyReply'
    /v1/config/create:
        post:
            tags:
                - SysConfig
            operationId: SysConfig_CreateSysConfig
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.CreateSysConfigRep'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.EmptySysConfigReply'
    /v1/config/list:
        post:
            tags:
                - SysConfig
            operationId: SysConfig_ListSysConfig
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.ListSysConfigRep'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.ListSysConfigReply'
...
    /v1/login:
        post:
            tags:
                - Auth
            operationId: Auth_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.LoginReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.LoginReply'
    /v1/logout:
        post:
            tags:
                - Auth
            operationId: Auth_Logout
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.LogoutReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.LogoutReply'
...
    /v1/register:
        post:
            tags:
                - Auth
            operationId: Auth_Register
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.RegisterReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.RegisterReply'
...
components:
    schemas:
...
        api.system.v1.ConfigByKeyReply:
            type: object
            properties:
                value:
                    type: string
        api.system.v1.ConfigReply:
            type: object
            properties:
                configId:
                    type: string
                configName:
                    type: string
                configKey:
                    type: string
                configValue:
                    type: string
                configType:
                    type: string
                remark:
                    type: string
                createdAt:
                    type: string
...
        api.system.v1.CreateSysConfigRep:
            type: object
            properties: {}
...
        api.system.v1.EmptySysConfigReply:
            type: object
            properties: {}
...
        api.system.v1.ListSysConfigRep:
            type: object
            properties:
                pageNum:
                    type: string
                    description: pageNum
                pageSize:
                    type: string
                    description: 页面大小
        api.system.v1.ListSysConfigReply:
            type: object
            properties:
                rows:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.system.v1.ConfigReply'
                total:
                    type: string
...
        api.system.v1.LoginReply:
            type: object
            properties:
                token:
                    type: string
        api.system.v1.LoginReq:
            type: object
            properties:
                username:
                    type: string
                password:
                    type: string
                uuid:
                    type: string
                code:
                    type: string
        api.system.v1.LogoutReply:
            type: object
            properties: {}
        api.system.v1.LogoutReq:
            type: object
            properties: {}
...
        api.system.v1.RegisterReply:
            type: object
            properties: {}
        api.system.v1.RegisterReq:
            type: object
            properties:
                username:
                    type: string
                nickname:
                    type: string
                mobile:
                    type: string
                password:
                    type: string
...